practice1/requirements.txt
fastapi
uvicorn
numpy
pillow
pytest
httpx
python-multipart